requirements1.txt
Flask
kafka-python
werkzeug
pybreaker
requests
pymongo
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-kafka-python
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-pymongo
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-logging

